First succesfull run

+++ networks.py
@@ -41,7 +41,7 @@ class CriticNetwork(nn.Module):
 
     def forward(self, state, action):
         x = F.relu(self.fc1(state))
-        x = torch.cat((x, action.float()), dim=1)
+        x = torch.cat((x, action), dim=1)
         x = F.relu(self.fc2(x))
         return self.fc3(x)
